# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ITIL v4 Change Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          description: Invalid credentials
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          description: Not authenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordBody" }
      responses:
        "204": { description: OK }
        "400": { description: Bad request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # ---------- USERS ----------
...
  /changes/{id}/approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Approval" }
  /approvals/{id}/vote:
    post:
      operationId: voteApproval
      tags: [approvals]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/VoteBody" } } }
      responses:
        "200":
          description: Voted
          content: { application/json: { schema: { $ref: "#/components/schemas/Approval" } } }

  # ---------- CAB ----------
...
  /audit-log:
    get:
      operationId: listAuditLog
      tags: [audit]
      parameters:
        - { in: query, name: actorId, schema: { type: integer } }
        - { in: query, name: action, schema: { type: string } }
        - { in: query, name: entityType, schema: { type: string } }
        - { in: query, name: entityId, schema: { type: integer } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: limit, schema: { type: integer } }
        - { in: query, name: offset, schema: { type: integer } }
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEntry" }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /audit-log/export:
    get:
      operationId: exportAuditLog
      tags: [audit]
      parameters:
        - { in: query, name: actorId, schema: { type: integer } }
        - { in: query, name: action, schema: { type: string } }
        - { in: query, name: entityType, schema: { type: string } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: CSV
          content:
            text/csv:
              schema: { type: string }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    LoginBody:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
    SessionUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        fullName: { type: string }
        source: { type: string, enum: [local, ldap] }
        roles:
          type: array
          items: { type: string }
        isAdmin: { type: boolean }
      required: [id, username, email, fullName, source, roles, isAdmin]
...
    Approval:
      type: object
      properties:
        id: { type: integer }
        changeId: { type: integer }
        roleKey: { type: string }
        roleName: { type: string }
        approverId: { type: ["integer", "null"] }
        approverName: { type: ["string", "null"] }
        decision: { type: string, enum: [pending, approved, rejected, abstain] }
        comment: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        viaDeputy: { type: boolean }
      required: [id, changeId, roleKey, roleName, decision, viaDeputy]
    VoteBody:
      type: object
      properties:
        decision: { type: string, enum: [approved, rejected, abstain] }
        comment: { type: ["string", "null"] }
      required: [decision]
...
    AuditEntry:
      type: object
      properties:
        id: { type: integer }
        timestamp: { type: string, format: date-time }
        actorId: { type: ["integer", "null"] }
        actorName: { type: string }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: ["integer", "null"] }
        summary: { type: string }
        ipAddress: { type: string }
        userAgent: { type: string }
        before: {}
        after: {}
      required: [id, timestamp, actorName, action, entityType, summary, ipAddress, userAgent]